Simplified rendering for 'normal' or 'stepped' pagination
